requirements.txt
amqp==5.1.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.14
cftime==1.6.2
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.3
GDAL==3.0.4
geographiclib==1.52
geojson==2.5.0
geomet==0.3.0
geopy==2.2.0
html2text==2020.1.16
idna==3.4
imageio==2.22.0
kiwisolver==1.4.4
kombu==5.2.4
landsatxplore==0.13.0
lxml==4.9.1
matplotlib==3.6.0
netCDF4==1.6.1
networkx==2.8.6
numpy==1.23.3
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
plotext==5.2.4
prompt-toolkit==3.0.31
pyhdf==0.10.5
pyparsing==3.0.9
pyproj==3.4.0
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.1
sentinelsat==1.1.1
Shapely==1.8.4
shtab==1.5.5
six==1.16.0
tifffile==2022.8.12
tqdm==4.64.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
